Add support for custom match functions
If b:match_function is defined, matchit will first call this function to
perform matching. This is useful for languages with an indentation-based
block structure (such as Python) or other complex matching requirements
that cannot be expressed with regular expression patterns.

If b:match_function doesn't return a result, we fall through to regular
matching (b:match_words, matchpairs, etc.).

Errors thrown by the b:match_function are fatal and won't continue. The
exception is printed when b:match_debug is enabled.